Fix problems with bean activation in different modes
