+++ fhir-server/src/main/java/au/csiro/pathling/io/ResourceReader.java
@@ -32,7 +32,6 @@ import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
 import org.hl7.fhir.r4.model.Enumerations.ResourceType;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
@@ -43,8 +42,7 @@ import org.springframework.stereotype.Component;
  * @author John Grimes
  */
 @Component
-@ConditionalOnMissingBean(ResourceReader.class)
-@Profile("core")
+@Profile("core & !ga4gh")
 @Slf4j
 public class ResourceReader {
