+++ fhir-server/src/main/java/au/csiro/pathling/security/PathlingAuthenticationConverter.java
@@ -9,7 +9,6 @@ package au.csiro.pathling.security;
 import java.util.Collection;
 import javax.annotation.Nonnull;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Profile;
 import org.springframework.core.convert.converter.Converter;
@@ -26,8 +25,7 @@ import org.springframework.stereotype.Component;
  */
 @Component
 @ConditionalOnProperty(prefix = "pathling", name = "auth.enabled", havingValue = "true")
-@ConditionalOnMissingBean(JwtAuthenticationConverter.class)
-@Profile("server")
+@Profile("server & !ga4gh")
 @Slf4j
 public class PathlingAuthenticationConverter extends JwtAuthenticationConverter {
