+++ fhir-server/src/main/java/au/csiro/pathling/security/ga4gh/PassportScope.java
@@ -9,7 +9,6 @@ package au.csiro.pathling.security.ga4gh;
 import java.util.HashMap;
 import java.util.Set;
 import org.hl7.fhir.r4.model.Enumerations.ResourceType;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.annotation.Scope;
 import org.springframework.context.annotation.ScopedProxyMode;
@@ -23,9 +22,7 @@ import org.springframework.web.context.WebApplicationContext;
  * @author John Grimes
  */
 @Component
-@ConditionalOnProperty(prefix = "pathling", name = {"auth.enabled", "auth.ga4gh-passports.enabled"},
-    havingValue = "true")
-@Profile("server")
+@Profile("server & ga4gh")
 @Scope(value = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
 public class PassportScope extends HashMap<ResourceType, Set<String>> {
